+++ front-end/src/app/components/elecciones/autoridad-edit/autoridad-edit.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Autoridad } from 'src/app/models/autoridad';
+import { AuthService } from 'src/app/services/auth.service';
 import { EleccionService } from 'src/app/services/eleccion.service';
 
 @Component({
@@ -19,7 +20,9 @@ export class AutoridadEditComponent implements OnInit {
   constructor(
     public eleccionService: EleccionService,
     public route: ActivatedRoute,
-    public router: Router
+    public router: Router,
+    public authService: AuthService,
+
   ) { }
 
   ngOnInit(): void {
@@ -54,15 +57,40 @@ export class AutoridadEditComponent implements OnInit {
   }
 
   editar() {
-    const estaSeguro = confirm('¿Está seguro que desea editar los datos de esta dignidad?');
-    if(estaSeguro)
-      this.eleccionService.updateDignidad(this.autoridad, this.lista, this.eleccion);
+    this.authService.auth.user.subscribe(
+      user => {
+        this.authService.getPermiso(user.email, 'Elecciones_edit').subscribe(
+          permiso => {
+            if (permiso.length > 0) {
+              const estaSeguro = confirm('¿Está seguro que desea editar los datos de esta dignidad?');
+              if (estaSeguro)
+                this.eleccionService.updateDignidad(this.autoridad, this.lista, this.eleccion);
+            }
+            else
+              alert('Usted no tiene permiso para realizar esa acción');
+          }
+        )
+      }
+    )
+
   }
 
   eliminar() {
-    const estaSeguro = confirm('¿Está seguro que desea elliminar esta dignidad?');
-    if(estaSeguro)
-      this.eleccionService.deleteDignidad(this.autoridad, this.lista, this.eleccion);
+    this.authService.auth.user.subscribe(
+      user => {
+        this.authService.getPermiso(user.email, 'Elecciones_edit').subscribe(
+          permiso => {
+            if (permiso.length > 0) {
+              const estaSeguro = confirm('¿Está seguro que desea elliminar esta dignidad?');
+              if (estaSeguro)
+                this.eleccionService.deleteDignidad(this.autoridad, this.lista, this.eleccion);
+            }
+            else
+              alert('Usted no tiene permiso para realizar esa acción');
+          }
+        )
+      }
+    )
   }
 
 }
