+++ front-end/src/app/components/elecciones/elecciones-lista/elecciones-lista.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Autoridad } from 'src/app/models/autoridad';
 import { Lista } from 'src/app/models/lista';
+import { AuthService } from 'src/app/services/auth.service';
 import { EleccionService } from 'src/app/services/eleccion.service';
 
 @Component({
@@ -11,21 +12,22 @@ import { EleccionService } from 'src/app/services/eleccion.service';
 })
 export class EleccionesListaComponent implements OnInit {
 
-  public lista: Lista = {nombre:""};
+  public lista: Lista = { nombre: "" };
   public listaParam: string;
   public eleccionParam: string;
-  public dignidades: Autoridad[] = []; 
+  public dignidades: Autoridad[] = [];
   public dignidadesNombres: string[] = [];
   public rutaDeRegreso: string = "/elecciones";
 
   constructor(
     private route: ActivatedRoute,
     private router: Router,
+    private authService: AuthService,
     private eleccionService: EleccionService
   ) { }
 
   ngOnInit(): void {
-    this.route.params.subscribe( params => {
+    this.route.params.subscribe(params => {
       this.eleccionParam = params['eleccion'];
       this.listaParam = params['lista'];
       this.rutaDeRegreso = "/elecciones/" + this.eleccionParam;
@@ -35,7 +37,7 @@ export class EleccionesListaComponent implements OnInit {
           this.eleccionService.getDignidadesDeLista(this.listaParam, this.eleccionParam).subscribe(
             dignidades => {
               this.dignidades = dignidades;
-              this.dignidadesNombres = dignidades.map(n=>n.Cargo);
+              this.dignidadesNombres = dignidades.map(n => n.Cargo);
             },
             error => {
               console.error(error);
@@ -47,9 +49,9 @@ export class EleccionesListaComponent implements OnInit {
         }
       )
     },
-    error => {
-      console.error(error);
-    })
+      error => {
+        console.error(error);
+      })
   }
 
   irADignidad(dignidad: string) {
@@ -57,16 +59,38 @@ export class EleccionesListaComponent implements OnInit {
   }
 
   newDignidad() {
-    this.router.navigate(['/elecciones', this.eleccionParam, this.listaParam, 'crear']);
+    this.authService.auth.user.subscribe(
+      user => {
+        this.authService.getPermiso(user.email, 'Elecciones_edit').subscribe(
+          permiso => {
+            if (permiso.length > 0) {
+              this.router.navigate(['/elecciones', this.eleccionParam, this.listaParam, 'crear']);
+            }
+            else
+              alert('Usted no tiene permiso para realizar esa acción');
+          }
+        )
+      }
+    )
   }
 
   deleteList() {
-    const estaSeguro = confirm('¿Está seguro que desea eliminar esta lista?');
-    if(estaSeguro) {
-      this.eleccionService.deleteLista(this.lista, this.eleccionParam);
-      this.router.navigate(['/elecciones', this.eleccionParam]);
-    }
+    this.authService.auth.user.subscribe(
+      user => {
+        this.authService.getPermiso(user.email, 'Elecciones_edit').subscribe(
+          permiso => {
+            if (permiso.length > 0) {
+              const estaSeguro = confirm('¿Está seguro que desea eliminar esta lista?');
+              if (estaSeguro) {
+                this.eleccionService.deleteLista(this.lista, this.eleccionParam);
+                this.router.navigate(['/elecciones', this.eleccionParam]);
+              }
+            }
+            else
+              alert('Usted no tiene permiso para realizar esa acción');
+          }
+        )
+      }
+    )
   }
-
-
 }
